Set up role/db for icinga web 2
Use chef vault for storing the password data. If using test
kitchen, then the databag in `test/fixtures` will be used instead.

If the vault or databag is missing, then a random password is
generated on each run. That causes Icinga Web 2 to have a new
DB password on each chef run.