Add local testing data bags
Can store JSON objects with details of applications. These details
are the app internal ID/name, the domains to configure, and secret
ENV variables for connecting to databases. Data bags can be encrypted
(unlike attributes) so it is safer to store secrets there.

Recipe has yet to be updated to read from these data bags.