Use new volume ID
As I am deploying to sarcee and not sarcee-staging, I needed a new
OpenStack mounted volume. It has a different volume ID, and the
first 20 characters of that ID are used here for matching the disk
ID in `/dev/disk/by-id`.

Only the first 20 chars are used because that is all that are shown
in the directory listing.