Add cookbook for hostsfile setup
Running the recipe from this cookbook will add hostsfile entries
for the other Rapid Access Cloud instances to this node. This allows
apps and tools on the node to refer to instances by name and not IP.
IPs can then be managed by this cookbook centrally, reducing needed
updates across cookbooks.