Add base for FROST installation
This separate recipe makes it an optional install. The recipe also
has its own stanza for Test Kitchen so it can be tested separately
from the Data Transloader, although eventually these stanzas will
be combined so that FROST is used for testing with the Data
Transloader locally (while a public STA is used in production).