Start up FROST automatically
By customizing the docker compose configuration file, we can change
the default values and add "restart: always" to have the services
auto-start when the instance boots.

The service is mapped to port 8080 on the host and is publically
accessible — this is probably fine for testing but should be limited
if deployed in production as write access is open.