Split ETL script into two files and setup DAG
DAG now calls download and upload in separate steps, so they can
be debugged and monitored separately. A DAG will be created for
each provider, starting with Environment Canada (other providers
not yet created).

It is very important to use sudo to run the ETL scripts as the
transloader system user, as otherwise the Ruby libraries won't be
found.