Manually install Ruby
The Ruby bundled with Ubuntu is 2.5.1, which is too old for some
methods used in the Data Transloader. While the Data Transloader is
supposed to work with Ruby 2.4+, it has only been tested lately with
Ruby 2.6 so this version must also be installed on the production
server.

Installing a newer version of Ruby also fixes the rubygems and bundler
issue.