Import more Environment Canada station data
This is a list of stations that are of interest to the Arctic
Community Sensor Web project, as the stations are in regions near
the communities. Some stations from the original list of stations
are no longer available as SWOB-ML files, and those stations have
been removed from this list.

All of the stations in this list are available in the SWOB-ML
station list CSV file from Environment Canada as of 2019-09-06. But
some of them do not have SWOB-ML files (possibly offline temporarily?).
Without the SWOB-ML files, the initial import is set to ignore
failure and any ETL on that station will be ignored when ran by
Airflow.

If the missing stations are brought back online, re-running the
chef-client will try to download the SWOB-ML for metadata again.