Install postgresql for Airflow
Airflow will be switched to use PostgreSQL instead of SQLite. This
improves performance and allows some DAGs to be run in parallel.

Note that the Data Transloader is not thread safe, as the file store
is not guaranteed to have safe transactions from multiple processes.
DAGs that run on the same station should NOT be ran in parallel.