Switch to DAG per Environment Canada station
This creates a separate DAG in Airflow for each Environment Canada
station defined in the node attributes. This was done to better
utilize parallelism as each station is safe to run separately with
a single worker. (The issue would be if multiple workers tried to
modify the same station at the same time).