Fix port redirect issue for Airflow
When accessing Airflow on port 80, it would proxy to this node on
port 80, which then proxies to Airflow on port 2080. This node's
nginx would then redirect the browser to port 2080 which doesn't
work. This commit adds a directive to not redirect to port 2080,
and instead use whatever port the browser originally used.