Add note about encrypted data bag usage
If the Chef Vault items don't exist, then Chef will try to decrypt
using the older encrypted data bag system. That system requires a
secret file to be installed on the node. This might actually be
installed with knife bootstrap if certain parameters are used; if
not, then the secret file has to be created manually.