Use different upload interval for Data Garrison
The data is published every 120 minutes, and running the DAG every
60 minutes is causing data to not be uploaded. If the DAG were
changed to run every 120 minutes, then data could be stale for longer
as the DAG interval may be considerably offset from the publishing
interval.

Instead, the DAG runs every 60 minutes but uses a custom DAG coded
for Data Garrison where uploads always use the last 24 hours instead
of the length of the schedule interval (which would be 60 minutes).