Decrease retries on DAGs
DAG runs are expected to fail, as sometimes the data is unavailable
for download. In those cases, Airflow *should not* attempt to re-run
the DAG as it will not work.

This commit removes the retry option for DAGs, which will also free
up the Airflow scheduler to run tasks faster.