Install GeoServer vector tiles extension
I was able to verify in the web interface that GeoWebCache offers
GeoJSON, TopoJSON, and Mapbox Vector Tiles as caching options.
(UTF Grid is already included in the base GeoServer installation.)

I removed the links for the GDAL and CSS plugins from the
attributes as the plugins are not needed.