Import EPSG:4326 OSM water data
This is imported *instead* of the lower-resolution EPSG:3857 polygons
as the EPSG:3857 does not re-project nicely to EPSG:4326 (neither
reprojection done by GeoServer nor by OGR into PostGIS fix the
errors that occur when rendering to vector tiles). Errors include
inability to render polygons properly close to north or south poles.

This higher-resolution dataset should be the same as the lower-res
dataset; I can probably use an SQL view in GeoServer to generate
lower-complexity shapes for vector tiles in the lower zoom ranges
using PostGIS functions.