Pass API Key to ETL upload scripts
Deploys the X-Api-Key used by API Gateway into the upload scripts,
which pass it to the ETL tool. This is necessary for write access
to our STA.

If the api key in the data bag or chef vault is nil or an empty
string, then no "X-Api-Key" header will be sent with upload tasks.