Upgrade to use Terraform 0.14
Provider version constraints are no longer used.
Found an odd case where the key says it is an id (`kms_key_id`) but
is expecting an ARN. If you pass an ID, then Terraform wants to
destroy and re-create the EC2 instance (wrong, undesirable). If you
pass an ARN, then everything is OK.