Downgrade to Ubuntu 18.04 and Chef Server 12
The past backups were made by version 12 and you gotta have that
version installed to restore a backup. And version 12 does not have
official support for Ubuntu Server 20.04 so I downgraded to use
Ubuntu Server 18.04.

Remember to run a reconfigure after a restore or the web server is
not there for some reason.

After doing this I was able to get Chef Infra Server running and
it communicates properly with `knife` on my local machine, as well
as `chef-client` on our other nodes.