Downgrade RAC to Terraform 0.12
This is necessary to fix some issues that occur when trying to
remove instances that no longer exist. If you try to do this on
0.13, Terraform complains about providers no longer existing. So
I reverted to Terraform 0.12 and removed the `edmonton` and
`barlow` instances as they no longer exist on RAC and were removed
using the web interface.