Set up additional HD for testing
This adds a 10GB disk for wiki storage, onto which ZFS will be
configured. In production a mounted volume will provide the same
role, but locally in test-kitchen we use a volume for VirtualBox.

I had to use a fork of the `kitchen-vagrant` gem as there is a bug
where adding an HD will not work because it tries to create the HD
file multiple times and fails; this fork fixes that.